Use float32 for reduced lda model dumper float16 is not supported by parquet reading data should still work, as msgpack is not deleted yet

+++ model_data_base/IO/LoaderDumper/reduced_lda_model.py
@@ -62,7 +62,7 @@ def dump(obj, savedir):
 
     try:
         mdb.setitem('st',
-                    Rm.st.round(decimals=2).astype('f2').reset_index(drop=True),
+                    Rm.st.round(decimals=2).astype(float).reset_index(drop=True),
                     dumper=pandas_to_parquet)
         del Rm.st
         del Rm.lda_values  # can be recalculated
